WIP: implement new features (bondStaked based quorum, cancel proposal, etc)

+++ contracts/interfaces/IBridge.sol
@@ -5,9 +5,9 @@ pragma experimental ABIEncoderV2;
 interface IBridge {
     function queuedTransactions(bytes32 txHash) external view returns (bool);
 
-    function queueTransaction(address target, uint value, string memory signature, bytes memory data, uint eta) external returns (bytes32);
+    function queueTransaction(address target, uint256 value, string memory signature, bytes memory data, uint256 eta) external returns (bytes32);
 
-    function cancelTransaction(address target, uint value, string memory signature, bytes memory data, uint eta) external;
+    function cancelTransaction(address target, uint256 value, string memory signature, bytes memory data, uint256 eta) external;
 
-    function executeTransaction(address target, uint value, string memory signature, bytes memory data, uint eta) external payable returns (bytes memory);
+    function executeTransaction(address target, uint256 value, string memory signature, bytes memory data, uint256 eta) external payable returns (bytes memory);
 }
